LUDOC-481 setup: update OSS/MDS RAM requirements
Update the MDS and OSS RAM requirements to reflect increased memory
requirements from the base OS for RHEL8, as well as the general
increase in the number of files that are being accessed by clients.

Also, document additional RAM for active-active DNE MDT failover
pairs, which wasn't taken into consideration when first written.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I39545466fbe39eb72a6745294f14f88e023ebbe5
Reviewed-on: https://review.whamcloud.com/41391
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>